Idk how the json files are being acessed

+++ bot.py
@@ -18,23 +18,9 @@ async def load():
         if filename.endswith('.py'):
             await bot.load_extension(f'cogs.{filename[:-3]}')
 
-async def create_guild_files():
-    print("create guild files")
-    async for guild in bot.fetch_guilds(limit=None):
-        print(guild + "1" + guild.id)
-        if not os.path.isfile(f'./data/{guild.id}.json'):
-            data = {
-                'quotes_chat': None,
-                'daily_quotes': None,
-                'history': []
-            }
-            with open(f'./data/{guild.id}.json', 'w') as f:
-                json.dump(data, f)
-
 async def run_discord_bot():
     async with bot:
         load_dotenv()
-        await create_guild_files()
         await load()
         await bot.start(os.getenv('TOKEN'))
 
@@ -48,51 +34,51 @@ async def ping(ctx):
 
 @bot.command()
 async def set(ctx):
-    with open(f'./data/{ctx.guild.id}.json', 'r') as f:
+    if not os.path.isfile(f'./data/{ctx.guild.id}.json'):
+            data = {
+                'quotes_chat': ctx.channel,
+                'daily_quotes': None,
+                'history': []
+            }
+            with open(f'./data/{ctx.guild.id}.json', 'w') as f:
+                json.dump(data, f)
+    with open(f'./data/{ctx.guild.id}.json', 'rw') as f:
         data = json.load(f)
-    data['quotes_chat'] = ctx.channel.id
-    temp_history = []
-    if data['history'] == []:
-            async for message in ctx.channel.history(limit=None):
+        data['quotes_chat'] = ctx.channel
+        if data['history'] == []:
+            async for message in data['quotes_chat'].history(limit=None):
                 if re.match(QUOTE_FORMAT, message.content) and message.author != bot.user:
-                    temp_history.append(message)
-    await ctx.send(f"#{data['quotes_chat']} has been set as the quotes chat.")
-    data['history'] = temp_history
-    with open(f'./data/{ctx.guild.id}.json', 'w') as f:
+                    data['history'].append(message)
+        await ctx.send(f"#{data['quotes_chat']} has been set as the quotes chat.")
         json.dump(data, f)
 
 @bot.command()
 async def rand(ctx):
-    with open(f'./data/{ctx.guild.id}.json', 'r') as f:
+    with open(f'./data/{ctx.guild.id}.json', 'rw') as f:
         data = json.load(f)
-    if data['quotes_chat'] == None:
-        await ctx.send('No quotes chat has been set.')
-    else:
-        temp_history = []
-        if data['history'] == []:
-            async for message in ctx.channel.history(limit=None):
-                if re.match(QUOTE_FORMAT, message.content) and message.author != bot.user:
-                    temp_history.append(message)
-        await ctx.send(f"#{data['quotes_chat']} has been set as the quotes chat.")
-        data['history'] = temp_history
-        random_quote = random.choice(data['history'])
-        speaker = random_quote.mentions[0].display_name
-        author = random_quote.author
-        quote = re.match(QUOTE_ONLY, random_quote.content)
-        date = random_quote.created_at.strftime("%d/%m/%Y")
+        if data['quotes_chat'] == None:
+            await ctx.send('No quotes chat has been set.')
+        else:
+            if data['history'] == []:
+                async for message in ctx.channel.history(limit=None):
+                    if re.match(QUOTE_FORMAT, message.content) and message.author != bot.user:
+                        data['history'].append(message)
+            await ctx.send(f"#{data['quotes_chat']} has been set as the quotes chat.")
+            random_quote = random.choice(data['history'])
+            speaker = random_quote.mentions[0].display_name
+            author = random_quote.author
+            quote = re.match(QUOTE_ONLY, random_quote.content)
+            date = random_quote.created_at.strftime("%d/%m/%Y")
 
         await ctx.send(f'@{speaker} said: {quote[0]} on {date}.')
-        with open(f'./data/{ctx.guild.id}.json', 'w') as f:
-            json.dump(data, f)
+        json.dump(data, f)
 
 @bot.command()
 async def daily(ctx):
-    with open(f'./data/{ctx.guild.id}.json', 'r') as f:
+    with open(f'./data/{ctx.guild.id}.json', 'rw') as f:
         data = json.load(f)
-    data['daily_quotes'] = ctx.channel.id
-    await ctx.send(f"#{data['data_quotes']} has been set as the daily quotes chat.")
-
-    with open(f'./data/{ctx.guild.id}.json', 'w') as f:
+        data['daily_quotes'] = ctx.channel.id
+        await ctx.send(f"#{data['data_quotes']} has been set as the daily quotes chat.")
         json.dump(data, f)
